Moving out code for Thermos, Thermos Observer, and the python 2 client. (#10)
The project will now solely focus on providing quality of life updates and maintenance for the scheduler component of Aurora.

Note that this will break end to end tests as currently scripted. It is possible to fix these scripts to run again using the source from the newly created repos for these components. In the meantime, the end to end tests provided by gorealis (https://github.com/aurora-scheduler/gorealis) may be used as a (temporary) substitute while those fixed.