Updating installation instructions
Updated installation instructions to reflect how a current install of Aurora Scheduler can be accomplished.

Since the project has decoupled the development of the scheduler from the development of the client and the worker node components, I have used the latest versions of the Aurora Scheduler binaries that exist for each platform as example installations. 

However, since the decoupling happened, there's no implicit guarantee that the scheduler will work with these components so I've also added a warning to test thoroughly.